Rename module font -> fonts
